# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/checkuserlogininfo": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "检查用户是否登录，并返回用户名，用户id。",
                "responses": {
                    "200": {
                        "description": "cookie",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api_models.RespData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api_models.UserCheckLoginRespAPI"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录，并返回cookie。",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户电话",
                        "name": "telephone",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cookie",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api_models.RespData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api_models.UserLoginRespAPI"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户电话",
                        "name": "telephone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.RespData"
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api_models.RespData": {
            "type": "object",
            "required": [
                "code",
                "data",
                "message",
                "success"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api_models.UserCheckLoginRespAPI": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api_models.UserLoginRespAPI": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}